api/requirements.txt
python-dotenv
psycopg2-binary
fastapi
uvicorn